# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DotNetSQL, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/OrderRecord": {
      "get": {
        "tags": [
          "DotNetSQL, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "GetOrderRecords",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OrderRecord"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DotNetSQL, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "CreateOrderRecord",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/OrderRecord/{id}": {
      "get": {
        "tags": [
          "DotNetSQL, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "GetOrderRecordById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "DotNetSQL, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "UpdateOrderRecord",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderRecord"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "DotNetSQL, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "operationId": "DeleteOrderRecord",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "OrderRecord": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "dateTime": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}